feat(code-quality): comprehensive clippy audit and fixes (ST-747)
- Fixed type complexity warnings by adding type aliases
- Fixed derivable impls warnings by using #[derive(Default)]
- Fixed new_without_default warnings by implementing Default trait
- Fixed field_reassign_with_default warnings by using struct update syntax
- Fixed reserve_after_initialization warnings by using Vec::with_capacity
- Fixed unnecessary_cast warnings by removing redundant casts
- Fixed unnecessary_map_or warnings by using is_ok_and method

All clippy warnings have been resolved and the code now passes clippy checks without errors.

+++ smoothtask-core/src/model/updater.rs
@@ -223,7 +223,7 @@ impl ModelUpdater {
         
         // Проверяем размер файла
         let content_length = response.content_length().unwrap_or(0);
-        if content_length != update_info.file_size as u64 {
+        if content_length != update_info.file_size {
             return Err(anyhow::anyhow!(
                 "Размер загруженного файла ({} байт) не совпадает с ожидаемым ({} байт)",
                 content_length,
